Remove getHumanChoice function
Completey removed getHumanChoice function, which
initated prompt for user input. Now the argument
is directly given from the button to the function
playRound. getComputerChoice stays the same.

Added the console.log of the Human Choice to the
playRound function so we still get a console message
of what the user chose.