Add functionality to play X rounds and pick winner
Added a new function so a winner is chosen after 4
rounds. The solution is not optimal and could be
improved. Had a problem that the user doesn't have
to press a button again so that the "Winner text"
is shown. Had a little workaround with a the new
endgame function.